+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Tests for "Seam carving".
+- Tests for "Rational number tree".
 
 ## [1.11.0] - 2022-07-31
 ### Added